Move mapObjects method to mapper-utils.js
